+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- `FmuInstance` is now generic over a `Borrow<FmuLibrary>` to allow the user to
+    managed `FmuLibrary` lifetime using Cell types.
+
 ## 0.3.0 - 2023-10-17
 
 ### Changed